requirements.txt
google-cloud-speech
wavio
scipy
sounddevice